add BitMarked service class to BlazorUI #10221
